demo_taller3/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.1
